Fix double promotion issues in chorus
This has sped up the code 3x. According to profiling, the chorus
effect has changed from taking 30% CPU time to just 10%.

+++ effects/chorus/src/chorus.cpp
@@ -7,13 +7,13 @@ namespace {
 
 float triangle(float phase)
 {
-    if(phase < M_PI)
+    if(phase < (float)M_PI)
     {
-        return 1 - 2 * phase / M_PI;
+        return 1 - 2 * phase / (float)M_PI;
     }
     else
     {
-        return - 3 + 2 * phase / M_PI;
+        return - 3 + 2 * phase / (float)M_PI;
     }
 }
 
@@ -45,12 +45,12 @@ void Chorus::process(float *samples, int sample_count)
 {
     for(int i = 0; i < sample_count; i++)
     {
-        float lfo = 0.5 * triangle(phase);
-        phase += modulation_rate / sample_rate * 2 * M_PI;
+        float lfo = 0.5f * triangle(phase);
+        phase += modulation_rate / sample_rate * 2 * (float)M_PI;
 
-        if(phase > 2 * M_PI)
+        if(phase > 2 * (float)M_PI)
         {
-            phase -= 2 * M_PI;
+            phase -= 2 * (float)M_PI;
         }
 
 #if 1
